backend/requirements-gmail.txt
# Gmail Integration Dependencies
# Install these alongside requirements.txt if you enable GMAIL integration
google-api-python-client>=2.115.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-auth>=2.27.0
